requirements.txt
streamlit>=1.36
tensorflow-cpu==2.20.*
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
